+++ libs/designer/src/lib/core/state/connection/connectionSelector.ts
@@ -19,6 +19,25 @@ import { useQuery } from '@tanstack/react-query';
 import { useSelector } from 'react-redux';
 import type { ConnectionsStoreState } from './connectionSlice';
 
+export const useConnectorOnly = (connectorId: string | undefined, enabled = true): UseQueryResult<Connector | undefined, unknown> => {
+  return useQuery(
+    ['apiOnly', { connectorId }],
+    async () => {
+      if (!connectorId) {
+        return null;
+      }
+      return await ConnectionService().getConnector(connectorId);
+    },
+    {
+      enabled: !!connectorId && enabled,
+      cacheTime: 1000 * 60 * 60 * 24,
+      refetchOnMount: false,
+      refetchOnWindowFocus: false,
+      refetchOnReconnect: false,
+    }
+  );
+};
+
 export const useConnector = (connectorId: string | undefined, enabled = true): UseQueryResult<Connector | undefined, unknown> => {
   const { data, ...rest }: any = useConnectorAndSwagger(connectorId, enabled);
   return { data: data?.connector, ...rest };