+++ libs/designer/src/lib/ui/panel/templatePanel/quickViewPanel/tabs/overviewTab.tsx
@@ -1,8 +1,10 @@
 import { isNullOrUndefined } from '@microsoft/logic-apps-shared';
-import type { RootState } from '../../../../../core/state/templates/store';
+import type { AppDispatch, RootState } from '../../../../../core/state/templates/store';
 import { useSelector } from 'react-redux';
 import type { IntlShape } from 'react-intl';
 import constants from '../../../../../common/constants';
+import { closePanel, openCreateWorkflowPanelView } from '../../../../../core/state/templates/panelSlice';
+import { clearTemplateDetails } from '../../../../../core/state/templates/templateSlice';
 
 export const OverviewPanel: React.FC = () => {
   const { manifest } = useSelector((state: RootState) => state.template);
@@ -15,7 +17,7 @@ export const OverviewPanel: React.FC = () => {
   );
 };
 
-export const overviewTab = (intl: IntlShape) => ({
+export const overviewTab = (intl: IntlShape, dispatch: AppDispatch) => ({
   id: constants.TEMPLATE_PANEL_TAB_NAMES.OVERVIEW,
   title: intl.formatMessage({
     defaultMessage: 'Overview',
@@ -30,5 +32,19 @@ export const overviewTab = (intl: IntlShape) => ({
   visible: true,
   content: <OverviewPanel />,
   order: 1,
-  icon: 'Info',
+  footerContent: {
+    primaryButtonText: intl.formatMessage({
+      defaultMessage: 'Create a workflow with this template',
+      id: 'wGkH/j',
+      description: 'Button text to create workflow from this template',
+    }),
+    primaryButtonOnClick: () => {
+      dispatch(openCreateWorkflowPanelView());
+    },
+    primaryButtonDisabled: false,
+    onClose: () => {
+      dispatch(closePanel());
+      dispatch(clearTemplateDetails());
+    },
+  },
 });