+++ libs/logic-apps-shared/src/utils/src/lib/models/template.ts
@@ -1,29 +1,33 @@
 export type SkuType = 'standard' | 'consumption';
-export type Kind = 'stateful' | 'stateless';
+export type WorkflowKindType = 'stateful' | 'stateless';
+export type ConnectorRuntimeType = 'inapp' | 'shared';
 
 export interface Manifest {
   title: string;
   description: string;
-  thumbnail?: string;
   skus: SkuType[];
-  kinds: Kind[];
+  kinds: WorkflowKindType[];
+  details: Record<string, string>;
+  tags?: string[];
   artifacts: Artifact[];
-  images?: string[];
+  images: Record<string, string>;
+  prerequisites?: string;
   parameters: Parameter[];
-  connections: Connection[];
+  connections: Record<string, Connection>;
 }
 
 export interface Artifact {
   type: string;
   file: string;
 }
-
 export interface Parameter {
   name: string;
   type: string;
   default?: string;
   description: string;
   required?: boolean;
+  displayName: string;
+  allowedValues?: { value: any; displayName: string }[];
 }
 
 export interface ParameterDefinition extends Parameter {
@@ -31,5 +35,6 @@ export interface ParameterDefinition extends Parameter {
 }
 
 export interface Connection {
-  id: string;
+  connectorId: string;
+  kind?: ConnectorRuntimeType;
 }